Update JNDI adapter to follow referrals by default (#182)
* Update JNDI adapter to follow referrals by default

I ran in to some issues while exploring OpenLDAP where
memberOf was not returning while using JNDI but was returning using
net-ldap and the ldapsearch command. I finally narrowed it down to JNDI
not following referrals by default.

Now this JNDI adapter will follow referrals by default like net-ldap
and the ruby ldap adapters do here. This can be changed with the
follow_referrals connection configurations option.

* Add test for reading memberOf attribute

* Add test_follow_referrals_option

* Update test_follow_referrals_option to test all three values

* Add dynamic groups test

This highlights the need for enabling the follow referrals configuration to load dynamic group
memberships.

* Update dockerfile and dockerignore

* Use dyngroup schema in slapd

* Move LDIF to enable dynamic groups to test/

* Remove needless GroupOfNames class

It's not related to this change.

* test: simplify

* Add support for follow_referrals per connect

* Enable no follow_referrals only for JRuby

Co-authored-by: Sutou Kouhei <kou@clear-code.com>