Use attribute_method_matchers instead of attribute_method_patterns with ActiveModel >= 7.1.0 (#195)
In ActiveModel 7.1.0 `attribute_method_patterns` was renamed to
`attribute_method_matchers`, see
https://github.com/rails/rails/pull/44367. As this method is used in
`AL::AttributeMethods#methods`, this caused errors when calling
`methods` on an instance of `AL::Base`.

This PR causes `methods` to always use `attribute_method_matchers` if it
exists. If not, `attribute_method_patterns` is called as a fallback to
support ActiveModel versions older than 7.1.0.

I also changed the creation of the attribute methods after that, as the
`AM::AttributeMethodMatcher` Class already has a method for that:
`method_name(attr)`.